fix recursive json serialization on Book and Reservation #1
