+++ book-api/src/main/java/com/rudy/bibliotheque/mbook/model/Reservation.java
@@ -1,8 +1,6 @@
 package com.rudy.bibliotheque.mbook.model;
 
-import com.fasterxml.jackson.annotation.JsonIdentityInfo;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import com.fasterxml.jackson.annotation.ObjectIdGenerators;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.rudy.bibliotheque.mbook.model.common.AbstractEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -15,10 +13,8 @@ import java.util.Date;
 @EqualsAndHashCode(callSuper = true)
 @Data
 @Entity
-@JsonIdentityInfo(
-        generator = ObjectIdGenerators.PropertyGenerator.class,
-        property = "id")
 public class Reservation extends AbstractEntity {
+    @JsonIgnoreProperties({"reservations", "ongoingReservations"})
     @ManyToOne
     private Book book;
 
@@ -36,6 +32,6 @@ public class Reservation extends AbstractEntity {
     private Date reservationEndDate;
 
     @Column(nullable = false, length = 100)
-    private Status status;
+    private ReservationStatus status;
 
 }
