+++ book-api/src/main/java/com/rudy/bibliotheque/mbook/model/ReservationStatus.java
@@ -2,17 +2,15 @@ package com.rudy.bibliotheque.mbook.model;
 
 import com.rudy.bibliotheque.mbook.util.Constant;
 
-public enum Status {
-    PENDING(Constant.STATUS_PENDING_CODE, Constant.STATUS_PENDING),
-    ONGOING(Constant.STATUS_ONGOING_CODE ,Constant.STATUS_ONGOING),
-    LATE(Constant.STATUS_LATE_CODE, Constant.STATUS_LATE),
+public enum ReservationStatus {
+    ONGOING(Constant.STATUS_ONGOING_CODE , Constant.STATUS_ONGOING),
     FINISHED(Constant.STATUS_FINISHED_CODE, Constant.STATUS_FINISHED),
     CANCELLED(Constant.STATUS_CANCELLED_CODE, Constant.STATUS_CANCELLED);
 
     private String code;
     private String name;
 
-    Status(String code, String name) {
+    ReservationStatus(String code, String name) {
         this.code = code;
         this.name = name;
     }
