+++ book-api/src/main/java/com/rudy/bibliotheque/mbook/model/converter/ReservationStatusConverter.java
@@ -1,16 +1,16 @@
 package com.rudy.bibliotheque.mbook.model.converter;
 
-import com.rudy.bibliotheque.mbook.model.Status;
+import com.rudy.bibliotheque.mbook.model.ReservationStatus;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 import java.util.stream.Stream;
 
 @Converter(autoApply = true)
-public class StatusConverter implements AttributeConverter<Status, String> {
+public class ReservationStatusConverter implements AttributeConverter<ReservationStatus, String> {
 
     @Override
-    public String convertToDatabaseColumn(Status status) {
+    public String convertToDatabaseColumn(ReservationStatus status) {
         if (status == null) {
             return null;
         }
@@ -18,12 +18,12 @@ public class StatusConverter implements AttributeConverter<Status, String> {
     }
 
     @Override
-    public Status convertToEntityAttribute(String code) {
+    public ReservationStatus convertToEntityAttribute(String code) {
         if (code == null) {
             return null;
         }
 
-        return Stream.of(Status.values())
+        return Stream.of(ReservationStatus.values())
                 .filter(c -> c.getCode().equals(code))
                 .findFirst()
                 .orElseThrow(IllegalArgumentException::new);
