+++ book-api/src/main/java/com/rudy/bibliotheque/mbook/search/ReservationSearch.java
@@ -10,4 +10,5 @@ import java.util.Set;
 public class ReservationSearch {
     private String userId;
     private String bookId;
+    private Set status;
 }
