+++ book-api/src/main/java/com/rudy/bibliotheque/mbook/web/controller/LoanController.java
@@ -33,19 +33,11 @@ public class LoanController {
 
     private Logger log = LoggerFactory.getLogger(this.getClass());
 
-    private ApplicationPropertiesConfig appProperties;
     private BorrowService borrowService;
-    private CopyService copyService;
-    private UserInfoService userInfoService;
-    private KeycloakRestTemplate keycloakRestTemplate;
 
     @Autowired
-    public LoanController(BorrowService borrowService, CopyService copyService, UserInfoService userInfoService, ApplicationPropertiesConfig appProperties, KeycloakRestTemplate keycloakRestTemplate) {
+    public LoanController(BorrowService borrowService) {
         this.borrowService = borrowService;
-        this.copyService = copyService;
-        this.appProperties = appProperties;
-        this.keycloakRestTemplate = keycloakRestTemplate;
-        this.userInfoService = userInfoService;
     }
 
     /**
@@ -61,55 +53,7 @@ public class LoanController {
     @PreAuthorize("hasRole('" + Constant.STAFF_ROLE_NAME + "')")
     @PostMapping
     public ResponseEntity<Borrow> saveALoan(@RequestBody LoanCreateDTO loanCreateDTO) {
-        log.info("Start method saveALoan");
-        if (loanCreateDTO.getUserId() == null) {
-            throw new InvalidIdException("User id has not been provided");
-        }
-
-        //link the copy
-        if (loanCreateDTO.getBookId() == null && loanCreateDTO.getCode() == null) {
-            throw new InvalidIdException("At least a book id or a copy code need to be provided");
-        }
-        Copy linkedCopy;
-        if (loanCreateDTO.getCode() != null) {
-            linkedCopy = copyService.getCopyById(loanCreateDTO.getCode());
-            if (linkedCopy == null) {
-                throw new NotFoundException("No copy available for book with id " + loanCreateDTO.getBookId());
-            }
-            if (linkedCopy.isBorrowed()) {
-                throw new ProhibitedActionException("Copy with id " + loanCreateDTO.getCode() + " is already borrowed");
-            }
-        } else {
-            linkedCopy = copyService.getAnAvailableCopyByBookId(loanCreateDTO.getBookId());
-            if (linkedCopy == null) {
-                throw new NotFoundException("No copy available for book with id " + loanCreateDTO.getBookId());
-            }
-        }
-
-        Borrow borrow = new Borrow();
-        borrow.setCopy(linkedCopy);
-        log.debug("Copy linked to the loan");
-
-        //Link the userInfos
-        UserInfo linkedUserInfo = userInfoService.getUserInfoById(loanCreateDTO.getUserId());
-        if (linkedUserInfo == null) {
-            ResponseEntity<UserInfo> response = keycloakRestTemplate.getForEntity("http://localhost:8080/auth/admin/realms/bibliotheque/users/" + loanCreateDTO.getUserId(), UserInfo.class);
-            if (response.getStatusCode() != HttpStatus.OK) {
-                if(response.getStatusCode() == HttpStatus.NOT_FOUND) {
-                    throw new NotFoundException("Can't find user with id " + loanCreateDTO.getUserId());
-                }
-                throw new CRUDIssueException("Can't get userInfos from server");
-            }
-        }
-        borrow.setUserInfo(linkedUserInfo);
-        log.debug("User linked to the loan");
-
-        log.debug("Update mandatory fields");
-
-        loanToPendingLogic(borrow);
-
-        Borrow newBorrow = borrowService.saveLoan(borrow);
-        if (newBorrow == null) throw new CRUDIssueException("Can't' create loan");
+        Borrow newBorrow = borrowService.saveANewLoan(loanCreateDTO);
 
         log.info("Method ended");
         return new ResponseEntity<>(newBorrow, HttpStatus.CREATED);
@@ -122,19 +66,7 @@ public class LoanController {
     @PreAuthorize("hasRole('"+ Constant.STAFF_ROLE_NAME +"')")
     @PutMapping(Constant.SLASH_ID_PATH + Constant.VALIDATE_PATH)
     public ResponseEntity<Borrow> validateALoan(@PathVariable Long id) {
-        Borrow borrow = borrowService.getLoanById(id);
-        if (borrow == null) {
-            throw new NotFoundException("Can't find loan with id " + id);
-        }
-        if (borrow.getLoanStartDate() != null) {
-            throw new ProhibitedActionException("This loan is already validated");
-        }
-
-        loanToOngoingLogic(borrow, appProperties.getLoanTimeInDays());
-
-        Borrow newBorrow = borrowService.saveLoan(borrow);
-
-        if (newBorrow == null) throw new CRUDIssueException("Can't' update loan");
+        borrowService.validateALoan(id);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
@@ -146,61 +78,9 @@ public class LoanController {
     @PreAuthorize("hasRole('"+ Constant.STAFF_ROLE_NAME +"')")
     @PutMapping(Constant.SLASH_ID_PATH + Constant.RETURNED_PATH)
     public ResponseEntity<Borrow> saveALoanReturn(@PathVariable Long id, @RequestBody Borrow borrowAdditionalInfos) {
-        Borrow borrow = borrowService.getLoanById(id);
-        if (borrow == null) {
-            throw new NotFoundException("Can't find loan with id " + id);
-        }
-        if (borrow.getLoanStartDate() == null) {
-            throw new ProhibitedActionException("This loan is in pending state");
-        }
-        if (borrow.getReturnedOn() != null) {
-            throw new ProhibitedActionException("This loan is already returned");
-        }
-
-        if (borrowAdditionalInfos.getStateAfterBorrow() != null) {
-            borrow.setStateAfterBorrow(borrowAdditionalInfos.getStateAfterBorrow());
-        } else {
-            borrow.setStateAfterBorrow(borrow.getStateBeforeBorrow());
-        }
-
-        loanToReturnedLogic(borrow);
-
-        Borrow newBorrow = borrowService.saveLoan(borrow);
-
-        if (newBorrow == null) throw new CRUDIssueException("Can't' update loan");
+        borrowService.saveALoanReturn(id, borrowAdditionalInfos);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-
-    public static void loanToPendingLogic(Borrow borrow) {
-        borrow.getCopy().setBorrowed(true);
-        borrow.getCopy().getBook().setAvailableCopyNumber(borrow.getCopy().getBook().getAvailableCopyNumber() - 1);
-
-        Date today = new Date();
-        borrow.setStateBeforeBorrow(borrow.getCopy().getCurrentState());
-        borrow.setLoanRequestDate(today);
-
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTime(today);
-        calendar.add(Calendar.DATE, 2);
-        borrow.setDeadlineToRetrieve(calendar.getTime());
-    }
-
-    public static void loanToOngoingLogic(Borrow borrow , Integer loanTimeInDays) {
-        Date today = new Date();
-        borrow.setLoanStartDate(today);
-
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTime(today);
-        calendar.add(Calendar.DATE, loanTimeInDays);
-        borrow.setLoanEndDate(calendar.getTime());
-    }
-    public static void loanToReturnedLogic(Borrow borrow) {
-        borrow.getCopy().setBorrowed(false);
-        borrow.getCopy().getBook().setAvailableCopyNumber(borrow.getCopy().getBook().getAvailableCopyNumber()+1);
-
-        Date today = new Date();
-        borrow.setReturnedOn(today);
     }
 
 }
